% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixa_processo.R
\name{baixar_processo}
\alias{baixar_processo}
\title{Baixa processo de primeira instância do TJDFT}
\usage{
baixar_processo(x, diretorio = ".")
}
\arguments{
\item{x}{número unificado do processo (20 dígitos) com ou sem separador.}

\item{diretorio}{Default para o atual.}
}
\value{
Baixa as páginas html.
}
\description{
Baixa processo de primeira instância do TJDFT
}
