% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixar_tjdft_cjsg.R
\name{baixar_tjdft_cjsg}
\alias{baixar_tjdft_cjsg}
\title{Baixar consulta jurisprudencial de segunda instancia do TJDFT}
\usage{
baixar_tjdft_cjsg(buscalivre, aspas = TRUE)
}
\arguments{
\item{buscalivre}{palavra ou expressao a ser buscada.}

\item{aspas}{TRUE. Coloca a expressao entre aspas}
}
\description{
Baixar consulta jurisprudencial de segunda instancia do TJDFT
}
\keyword{Courts,}
\keyword{Decisions,}
\keyword{Jurimetry,}
\keyword{Webscraping}
