% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixar_tjdft_cpopg.R
\name{baixar_tjdft_cpopg}
\alias{baixar_tjdft_cpopg}
\title{Baixa processo de primeira instância do TJDFT}
\usage{
baixar_tjdft_cpopg(x, diretorio = ".")
}
\arguments{
\item{x}{número unificado do processo (20 dígitos) com ou sem separador.}

\item{diretorio}{Default para o atual.}
}
\value{
Baixa as páginas html.
}
\description{
Baixa processo de primeira instância do TJDFT
}
