% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixar_tjdft_decisao.R
\name{baixar_tjdft_decisao}
\alias{baixar_tjdft_decisao}
\title{Baixar inteiro teor da consulta jurisprudencia do TJDFT}
\usage{
baixar_tjdft_decisao(registro)
}
\arguments{
\item{registro}{String of registro's numbers to search for. You get this number by using the tjdf_meta function.}
}
\description{
Baixa inteiro teor dos acórdaos
}
\keyword{Courts,}
\keyword{Decisions,}
\keyword{Jurimetry,}
\keyword{Webscraping}
