% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inteiro_teor_df.R
\name{inteiro_teor_df}
\alias{inteiro_teor_df}
\title{Function inteiro_teor_df}
\usage{
inteiro_teor_df(registro)
}
\arguments{
\item{registro}{número de registro obtido com a função tjdf_meta.}
}
\description{
Esta função baixa os pdfs com o inteiro teor dos acórdãos do Tribunal de Justiça do Distrito Federal
}
\examples{
registro<-1005250
inteiro_teor_df(registro)
}
\keyword{Decisão}
\keyword{Inteiro}
\keyword{Judicial,}
\keyword{Jurisprudência,}
\keyword{Justiça,}
\keyword{Teor}
\keyword{Tribunal}
\keyword{de}
