% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_tjdft_cpopg.R
\name{ler_tjdft_cpopg}
\alias{ler_tjdft_cpopg}
\title{Lê dados processuais}
\usage{
ler_tjdft_cpopg(diretorio = ".", plano = NULL)
}
\arguments{
\item{diretorio}{onde se encontram os htmls}

\item{plano}{check `future::plan`}
}
\value{
tibble com as informações processuais.
}
\description{
Lê dados processuais
}
