% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjdf_meta.R
\name{tjdf_meta}
\alias{tjdf_meta}
\title{Function tjdf_meta}
\usage{
tjdf_meta(url, form)
}
\arguments{
\item{url}{A forma mais conveniente de obter o url é fazer a busca jurisprudencial
no TJDF, copiar a url da barra de endereço e transformá-la em uma string do R.}

\item{form}{Lista com informações para o body.}
}
\description{
Esta função extrai os metadados das decisões judiciais do Tribunal de Justiça do Distrito Federal
}
\examples{
tjdf_meta(url,form)
}
\keyword{Decisão}
\keyword{Judicial,}
\keyword{Jurisprudência,}
\keyword{Justiça,}
\keyword{Tribunal}
\keyword{Webscraping}
\keyword{de}
