% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inteiro_teor_df.R
\name{inteiro_teor_df}
\alias{inteiro_teor_df}
\title{Function inteiro_teor_df}
\usage{
inteiro_teor_df(registro)
}
\arguments{
\item{registro}{String of registro's numbers to search for. You get this number by using the tjdf_meta function.}
}
\description{
This function downloads the pdfs from Brazilian Distrito Federal second degreee decisions.
}
\examples{
registro<-1005250
inteiro_teor_df(registro)
}
\keyword{Courts,}
\keyword{Decisions,}
\keyword{Jurimetry,}
\keyword{Webscraping}
