% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjdf_meta.R
\name{tjdf_meta}
\alias{tjdf_meta}
\title{Function tjdf_meta}
\usage{
tjdf_meta(BuscaLivre, quoted = TRUE)
}
\arguments{
\item{BuscaLivre}{a word or an expression to searched for.}

\item{quoted}{TRUE. It will treat an expression between quotes.}
}
\description{
This function retrieves the metadata related to the high court decisions according to a search.
}
\examples{
tjdf_meta("dano moral coletivo", quoted=TRUE)
}
\keyword{Courts,}
\keyword{Decisions,}
\keyword{Jurimetry,}
\keyword{Webscraping}
